+++ TimberCommons/package/CHANGELOG.md
@@ -1,5 +1,11 @@
+# v1.1 (pre-release)
+* [Fix #4] In rare cases the game crashes when valve gets built.
+* Add `deDE` localization from @juf0816.
+* Update `ruRU` localization.
+* Add `Particle System` property to the `WaterValve` component.
+
 # v1.0 (25 Jul 2023)
 * Initial version of the mod.
-* `DirectWaterServiceAccessor` class to handle custom water mvoers.
+* `DirectWaterServiceAccessor` class to handle custom water movers.
 * `WaterValve` component for one-way water valve that mover water based on the levels at the input and
   output.